Use a coverage.py dependency version compatible with covimerage
The version of coverage required by covimerage is not compatible with
the current coverage version. Ideally, rather than downgrading it for
our python tests, we would just use a different requirements file for
vimscript tests, and remove the plugin from .coveragerc.

But this isn't possible as there's no way to tell `codecov` to use a
different coveragerc, and therefore the covimerage plugin has to be
loaded no matter what.

So we just restrict the coverage version to make covimerage work.