Vim coverage: Combine coverage from all tests
Previously we were running :profile start <file> on every test script
run. This was rewriting the <file> for each run and then we were only
reporting on the _last_ such run of vim (i.e. each script).

Instead, we call covimerage for each script and pass --append so that
the resulting .coverage.vim contains all of the covered code, not just
the last script worth.