Close completion menu when no candidates
Close the completion menu (get out of ctrl-x mode) when the set of
completions we generate is empty. This gets the user out of ctrl-x mode
and allows insert-mode mappings to work again.

Fixing this caused a regression in the completedone handing, which
highlighted another issue: we were not correctly identifying
non-user-entered insert-mode changes during TextChangedP. We were
relying on CompleteDone, but this is triggered too late or not at all.
The only reliable way when tabbing though elements of the pum is the
CompleteChanged event, which when combined with the
v:event.completed_item key tells us that the change was due to a user
changing the pum selection.

Fixing the above resolved the regression in handling completion fixits
and allows us to close the pum at the right time, but doing only one
leaves one or other broken, so we have to guard the new API only to
users with the CompleteChanged event available.

It was initially thought that this meant we could skip calling
s:Complete in the TextChangeInsertMode handler, but we actually want to
call it anyway to ensure that it's _our_ completions that are shown, not
any vim-based filtering.