Add auto-hover behaviour (on by default)
This adds a hover feature, enabled by default on the CursorHold
autocommand (default: 4s), along with a <Plug> mapping to manually
trigger it.

We pick the subcommand to use based on preference order: GetHover,
GetDoc, GetType and remember the choice to avoid too many blocking
calls.

We enable syntax highlighting in the popup like in the signature help
popup, using the syntax of the current buffer. This isn't perfect
because some hover responses are markdown, though we tend to try and
hide that in our GetDoc and GetType. So we use `markdown` for GetHover
and `&syntax` for GetDoc/GetType.

Technically the response fromn the server is either plaintext or
markdown. We could actually use markdown.<current lang> in case that
gives better results with a good markdown syntax file, though that's not
usually the case. We could change ycmd to always request markdown and
insist users set g:markdown_fenced_languages, but that's a bit of a
stretch too.

Does not work in neovim due to using a popup.