Slightly improve the syntax heuristic for hover
We use this heuristic:

* For GetHover - use markdown
* For GetDoc - use nothing
* For GetType - use the current file syntax

You can also (undocumentedly) set b:ycm_hover to a dict with 'command'
and 'syntax' keys to manually override this if you so choose.

This also improves the tests - they now mock the python layer so that we
actually test the vimscript properly and cover the different cases.

At the same time, close the popup on 'word' rather than the default
'WORD' cursor movement.