Improve test time to run
Minimise the delays, and allow customising the delay when running
locally (e.g. on faster hardware). Default to 2 seconds, but for the
diagnostics tests, which require jdt.ls to start up, remain at 7
seconds.

This change also fixes a big instability with the signature_help tests
where we weren't waiting at all in the multiple filetypes scenario.