Imrpove signature help test stability
We need to actually wait for the signature_help_available responses to
have been received and for the result to not be PENDING.
Previously we were really only waiting for the response, but if it was
PENDING, we still could fail the test. Now we wait until it's not
PENDING for any native filetype.