Use `'close': 'click'` behavior for hover popups
The behavior used to be `'close': 'button'` which shows an 'X' in the top right of the popup. It's some times not clear that the 'X' is a button. The popup disappears when the cursor moves and the hover command now has toggle behavior so it seems less necessary. The new `'close': 'click'` behavior closes the popup when clicking anywhere in the popup, so the mouse can still be used.