Maximise displayed text for hover popup
Previously we relied on popup_atcursor to display the popup in a
sensible place. However, by default 'wrap' is enabled, which prevents
vim from shifting the popup to the left to show it on the screen (even
if you set minwidth). As a result it was possible for the popup to
display a very small vertical section when triggered close to the
right-hand screen edge.

The solution is to enable wrapping, which lets vim shift the popup. The
edge-case is that we _do_ want wrapping if there are lines that are
likely to be wrapped. Therefore in that case (when any line is > some
maximum width), we just force the popup to be placed at column 1 and
enable wrapping.

The maximum width for this wrapping is calculated as the screen width in
columns minus the padding we add in the x-axis. THis is a bit hacky as
it doesn't work properly for multi-byte characters, though the
calculation is likely to be _conservative_ in taht case, so it's
probably ok.