Fix signature help traceback when using preview popup
When we made YCM parse buffers with no filetype, this ended up including
the buffers within things like popups. Unfortuantely this leads to a
problem that we call "ClearSignatureHelp" from within
"FileReadyToParse". If the "FileReadyToParse" in question was within a
popup, we ge an error "Not valid in popup window" (referring to a call
to popup_close()) and vim gets in a very bad state.

it's not obvious how to solve this in the general case, but for now as
completion in non-filetype buffers is new and rare, we Blacklist enpty
buffers by default to work around the bug.

We allow users to explicitly enable it (and suffer this bug!) by
explicitly whiteliisting it again in ycm_filetype_whitelist.

Added a test which shows that it works when ycm_nofiletyp is
blacklisted, and another test which exercises the (still) broken
behaviour, when ycm_nofiletype is whitelisted.