Interractive symbol finder
We provide "GoToSymbol" and "GoToDocumentOutline" using a
"narrow-down-while-typing" approach similar to the insert completion
popup.

Thre are 2 modes:

* "Workspace" search (GoToSymbol) relies on the server (i.e. the LSP
  server) doing the filtering of the query. This is because none of the
  LSP servers I tested would return anything with query = ''. Therefore
  we can't easily use YCM's "filter and sort candidates" which would be
  preferable for consistency of "fuzzy" matching

* "Document" search (GoToDocumentOutline) uses YCM's filter and sort
  candidates because there is no server-side (i.e. LSP server) filtering
  for this request - it's all or nothing.

After selecting an item, we jump to that item and if there are more than
1, we populate and open the quick fix window with any matches.

We use a prompt buffer for the input, and use the TextChanged events to
update the current query, which is drawn in a massive popup taking up
the entire screen. We use the WinLeave event on the window containing
the prompt buffer also to cancel, along with ctrl-c. Escape remains the
key to exit insert mode in the buffer.

While the popup is open, we catch and handle some keys:

* C-j, Down, C-n - select the next item
* C-k, Up, C-p - select the previous item
* Enter - jump to the selected item
* C-c cancel/dismiss the popup
* Leave the prompt buffer - cancel/dismiss the popup.

No neovim support.

As some servers are quite slow (i'm looking at you Jedi), we have a
little spinner that displays while a request is outstanding.

A large comment in finder.vim describes the detail of the
implementation.