Fix auto imports when resolving completions on-demand
When we resolve completion items on-demand, we re-use the latest
completion request object to do the resolve request. This not only makes
sure that we only have one such request outstanding, but also that the
original request data items and query never change (as required by the
ycmd protocol).

The problem with this is that the complete done handling is implemented
by the CompletionRequest object. If we do a resolve, then we have lost
the completion request object to do the complete done handling. So to
resolve this we just delegate that handling back to the original
completion request when the latest "completion" request is actually a
resolve request.