Fix communication with WSGI server
The WSGI server won't accept:
  GET http://<addr>/ready HTTP/1.1
it requires:
  GET /ready HTTP/1.1

This is strictly conforming, as the HTTP RFC says:

> The most common form of Request-URI is that used to identify a
resource on an origin server or gateway. In this case the absolute
path of the URI MUST be transmitted (see section 3.2.1, abs_path) as
the Request-URI, and the network location of the URI (authority) MUST
be transmitted in a Host header field. For example, a client wishing
to retrieve the resource above directly from the origin server would
create a TCP connection to port 80 of the host "www.w3.org" and send
the lines:

   GET /pub/WWW/TheProject.html HTTP/1.1
   Host: www.w3.org

So as we're already sending the Host: header, just strip off the URL
prefix.