Make diagnostic highlighting in buffers without filepath work
Currently this is the only thing that doesn't work for this kind of
files.

Because we don't have a file path, we use `$PWD/$BUFNR` as file path in
requests. This mostly works, but for diagnostics it means we can't find
the file corresponding to the diagnostics we have received. See
vimsupport.py line 187 for reference.

This PR introduces some heuristics to make those work in most cases:

1. If the filepath we get from diags has a corresponding buffer number
   already, just use that. This covers already open files.
2. If the filepath doesn't exist on the local storage, and the tail of
   the path is a numeric string, we assume the weird path came from
   `$PWD/$BUFNR` magic.
3. Else we fall back to the current behaviour - either the path is an
   actual file path, or is not numeric. Either way, creating a new
   buffer seems reasonable.

Note that this fails under the following circumstances:

1. User has a file called `1` in the working directory.
2. User opens vim without any arguments.
3. YCM sends `OnBufferVisit` with `$PWD/1`.
4. ycmd responds with some diagnostics.
5. YCM mistakes `$PWD/1` (the buffer's fake path) with the actual file
   in the project.

A more robust solution could be something like a global mapping from
random strings to buffer numbers and using those generated strings as
file paths. I just didn't think it would be worth the effort.

Honestly, not sure if even this is worth the effort, but hey... there
was clearly an intention to keep this category of files working.