Do not call complete() if the cursor changed position since the completion request
This can happen if there's another async plugin that can change cursor
position, like matchup. In that case, it's a race condition between YCM
completing correctly and the plugin movin the cursor some place else.

The workaround here is to make sure the cursor position at request is
the same as cursor position at `complete()` time. As a result:

1. Erratic cursor movement won't mess up the buffer content and get in
   the way of typing.
2. Occasionally, YCM won't call `complete()` in a situation where a user
   might expect it to, but that's a direct consequence of `1`.
3. YCM will have spuroius "identifier complete" actions which means that
   "half" of a word might end up in identifier completer's database and
   subsequently suggested as a candidate. This is only temporary and the
   whole thing is eventually-consistent.