Do not assume current buffer in vimsupport.LineAndColumnNumbersClamped()
Before the big diagnostics refactor, we only served diagnostics for the
current buffer. That allowed `vimsupport.LineAndColumnNumbersClamped()`
to assume `vim.current.buffer` was always correct.

That is no longer the case - we were clamping offsets to, potentially,
the contents of a wrong buffer. More specifically, if

1. Current buffer is not the same buffer we are updating diagnostics
   for.
2. Current buffer does not have line number of the target buffer
   diagnostics.

Then we end up wrongfully "moving" the line of the diagnostic. On top of
that, if the target buffer's "moved" line is shorter than the start column of
the diagnostic, then we're still out of bounds.