Avoid KeyError when a different plugin's property type gets deleted
`vimsupport.GetTextProperties()` goes over all the properties in a
buffer. Then it filters for those whose type starts with `Ycm`.

However, if a plugin ever deletes its property type but leaves around a
property of that type, `prop_list()` will not contain the property type
for that property. This is where we previously encounter a rather nasty
`KeyError`.

I really wish vim had a nicer API for listing properties.