Move ycm_state.UpdateMatches() to s:OnBufferEnter()
This change does away with s:UpdateMatches(), for the sake of
performance in large files.

Previously, we would still try to update diagnostics in large files,
which would result in YCM calling a vim API for every line in that large
file.

The obvious solution - just call `s:AllowedToCompleteInCurrentBuffer()`
in `s:UpdateMatches()` - doesn't work, because it make
`s:OnBufferEnter()` return early as well, because
`s:VisitedBufferRequiresReparse() == false`. That makes
`s:OnBufferEnter()` skip the rest of the setup of completion.

Yes, this means that vimscript API for getting the text properties is
very inefficient.