Correctly handle multiline diagnostics in diagnostic_interface
While there doesn't seem to be a bug here (#3993 not withstanding),
there might be some improvement possible.

Previously, when we would attempt `props_to_remove.remove( multiline_prop )`,
it was wrong 100% of the time. This PR attempts to fix that.

The thing to keep in mind is that, on vim, `prop_list()` never returns
multiline properties. Instead, it splits multiline properties into
multiple single line properties. To figure out if should skip adding a
multiline property, we also need to split it by lines.

Open questions:

1. Is this a net negative in the average case, when it comes to
   performance?
2. Since we need to split props into single line ones, should we just
   always deal with single line props?
3. Does this mess anything up regarding neovim?