Implement manual toggling of signature help popup visibility
The signature help is awesome when you need it, but can get in the way,
when one needs to see the line above cursor.

With this commit, and an additional binding, it is possible to
temporarily hide the popup.

One caveat is that, if there were no popup to begin with and user starts
to edit arguments somewhere between triggers, the user won't be able to
spawn a popup with this binding.