Fix traceback with invalid diagnostic ranges
SOmetimes, for some reason, the legacy libclang completer returns ranges
where the end of the diagnostic range is 0,0 (or otherwise invalid).
It's not clear why that is, but it's coming from libclang and there's
not much we can do about it.

ANyway in the past we would have just crated a diagnostic "match" that
would have never matched anything. Now, we're trying to create a text
property spanning that range, which causes a vim error and ugly
traceback.

Solution is to simply validate ranges and drop invalid ones.