Reduce stutter by halving the redraws when typing
Previously we would trigger completion (call complete()) on every
keystroke even if nothing had changed in the result set. The purpose of
this with the old completion system (using feedkeys) was to reduce
flicker - we closed the completion menu on every keystroke, then
re-opened it.

With the new completion system, that's not necessary as we don't close
the pum on every keystroke - only when the result set is emtpy. However,
we still called complete() on every keystroke to ensure that it's always
our own filtering etc. that gets applied.

The problem with this is that this causes a _lot_ of redraw. Every call
to complete() seems to involve a very expensive draw call, particularly
in languages with complex regex-based syntax highlighting. Profiling vim
suggests that this change roughly halves the time spend redrawing -
which is to be expected.

I think that the extremely slow redraw is a bug in either the syntax
file or vim itslef, frankly, but this is quite difficult to pin down.
For now, I think it's better to have occasionally-incorrect filtering
when deleting characters to frequently-extremely-laggy typing.