diag virtual-text: Fix disabling of diagnostic updates in insert mode
This is more of an issue now that we echo diags inline - previously we
were updating the echo'd (or virtual text) diagnostic even if diag
updates were disabled in insert mode. This was annoying.

OnCursorMoved is called for both normal and insert mode, so just don't
update in insert mode.

Also clear virtual text when going into insert moode if updates are
disabled. This feels really nice to work with, by comparison with all
the noise when it's enabled, so recommend that setting in the docs.
Perhaps we can in future make it the default.