Update youcompleteme.vim
line 171: correct the spelling of 'enabled'.
line 1228: avoid showing line numbers in the info popup beside the completion popup menu in insert mode since line numbers often take too much space.
line 1591-1599: If you scroll down the hover in normal mode, it will automatically scroll back to its top after &updatetime's CursorHold. This change fixes the problem.
line 1630: add a border to the hover in normal mode to make it more distinct.
line 1634: set the maxheight of the hover in normal mode to fit the window size (i.e. keep the hover from going out of the window's boundary). Originally, the hover cannot be scrolled because it always fits its contents (some of its contents that are out of the window's boundary will become invisible). This change fixes the problem.