Provide a way to manually trigger YCM's completions
When automatic triggering is disabled, we now set the completefunc, so
the user can use C-x C-u. We also provide a plug mapping that does the
completion as users may not be comfortable with c-x c-u.

As this is a very unusual use-case for YCM we sweep under the rug that
most users probably want to use <Tab> for both completion and "next". Or
perhaps want to use C-p in the same way. For now a unique mapping works.