Reduce overhead of semantic highlighting (part 1)
Re-use the "optimal range" drawing used by inlay hints, i.e. only
re-request if the new visible range is not covered by the previously
requested visible range.

Essentially, semantic highlighting and inlay hints now use the (almost)
exact same code.

Still TODO (for both): record the _actual_ received range and compare
against that. The point is that we should only redraw if the new range
isn't within the last _drawn_ range. This is important for semantic
highlighting because servers might not support range highlighting and we
should therefore paint the whole buffer and only re-do that when it
changes.