Grow the requested range based on what we receive
This means that if the servers return a wider range of semantic tokens
or inlay hints than we requested, we don't constantly re-request them
whenever the viewport moves.

Importantly, clangd doesn't support range tokens, so we previously
requested and re-parsed, deleted and re-added every single token on
every single scroll event. This meant that scrolling was very slow and
lots of wasted effort. This should help reduce that. We should
essentially re-use the tokens on scolling and pretty much never
re-request them unless the buffer changes or we have a
file-ready-to-parse.