Fix custom text properties for semantic highlighting
We were supposed to allow users to set highlight groups for any token
types that we don't understand, as servers can basically return anything
they like. This didn't work and we just always complained about it.

Now we trap the "text property type doesn't exist" error and warn then.
This allows users to specify the missing ones to stfu the error. Also
improved the message to point to the help.