Make switching to semantic completer work after select ID completer candidate
To play nice with some other plugins, we only trigger completion if the
cursor has not moved between making a request and displaying the
results.

Previous oversight led to being unable to switch to semantic completion
after selecting an identifier completer candidate that changes the
cursor position.

The fix is to update the last cursor position from which the completion
request was made when the user hits <C-Space>