Handle multiple identical diagnostics in YcmShowDetailedDiagnostic popup
If a line for which detailed diagnostics are requested contains
more than a single diagnostic with the same message, YCM will try
`options.pop( 'col' )` more than once.

We do not need to really iterate through the diagnostics once we have
found the first one that is a match. If there's only one diag with the
matching message, looping beyond that diagnostic is just a waste of
time. If there's multiple diagnostics with the same message, it does not
matter which one we display in the popup.

Hence, the added break at the end of the loop.