Style the hierarchy popup like the finder popup
This basically involves moving the properties mapping to its own place,
and creating text properties to overlap the various parts of the popup
columns. Style matches, and feels correct.

Fiddly part is that we have to (for some reason) set cursorline every
time we move the cursor, but ain't nobody gonna be able to explain why,
and why that only necessary after launching the finder popup window
once. Some wierd vim bug is my guess.