Trim leading/trailing whitespace from description
This makes the columns seem less broken, particularly for leading
whitespace. This is particularly problematic due to our use of tabstop:
if the leading whitespace was a tab, it would go nuts.

It would still go nuts for any other literal tab in the description.
Perhaps we should fix that too.