Fix re-rooting of call/type hierarchies
Previously, when changing direction of hierarchy and re-rooting the
hierarchy tree, we would retain the old node as the new root.
That works most of the time, but fails in the face of multiple
inheritance and switching to supertypes. In that, problematic, case, the
resulting hierarchy will be missing some supertypes.
Worse, if a call hierarchy contains a node with multiple locations,
re-rooting to that node behaved in a way that defies explanation.

The solution is to throw away everything and start a new hierarchy
completely every time a user requests re-rooting.

This trades some performance for correctness, but this is not at all
performance sensitive.