Fix sending requests for a location in an unloaded buffer
In order to send a ycmd request for an unloaded buffer one must first...
load the buffer.
That is the only way to reliably determine the buffer's filetype.
However, we do not want to switch to the buffer, so everything needs to
be done "in the background".

`:badd` adds a buffer as unloaded, if it were not present before.
`bufload()` loads a buffer that has previously not been loaded.