jdt.ls: properly handle URI by showing contents
Thanks to commit ad64e63b (jdt.ls: support jumping to `jdt://` URIs with GoTo,
2025-07-07) on the ycmd repository, it has the capability to properly handle
`jdt://` URIs that are predominantly returned when a user executes the
`:YcmCompleter GoTo` command or some variation of it.

Introduce _GetClassNameFromJdtURI() which gets the classname from a JDT URI
(e.g. Member.class) using regular expression magic so that ycm displays it on
the status bar of the temporary buffer which is made to see the contents of that
class file.

On _HandleGotoResponse(), add a new condition which checks to see if there
is a key called 'jdt_contents' in the JSON response that ycmd provides.

If that condition is true, create a temporary readonly buffer, set the buffer
name to the file name that is received at an earlier stage from the 'uri' key,
and set the contents of the buffer to match what is received from the key
'jdt_contents'.

Finally, set the syntax highlighting to 'java' and jump the cursor to the line
and column specified from the JSON (specifically found in the 'range' key).

Note that the temporary buffer made does _NOT_ get its contents from any file -
no new temporary file is created or deleted during this process, everything is
handled inside vim.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>